% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes_measure_data}
\alias{nhanes_measure_data}
\title{NHANES PA data}
\format{
A data frame with 87619 rows and 9 variables:
\describe{
\item{SEQN}{ID variable}
\item{data_release_cycle}{wave/data release cycle}
\item{cat_age}{age category}
\item{gender}{sex/gender designation}
\item{wtmec4yr_adj_norm}{normalized weight for surveys}
\item{masked_variance_pseudo_psu}{PSU - sampling unit}
\item{masked_variance_pseudo_stratum}{sampling stratum}
\item{measure}{measure that was calculated}
\item{value}{value for the measure}
}
}
\source{
NHANES 2011-2012 and 2013-2014 accelerometer data.
}
\usage{
nhanes_measure_data
}
\description{
NHANES PA data
}
\keyword{datasets}
